Move levenshtein, uuid, overlay to datafusion-functions
