+++ datafusion/functions/Cargo.toml
@@ -75,6 +75,8 @@ log = { workspace = true }
 md-5 = { version = "^0.10.0", optional = true }
 regex = { version = "1.8", optional = true }
 sha2 = { version = "^0.10.1", optional = true }
+uuid = { version = "1.7", features = ["v4"] }
+
 [dev-dependencies]
 criterion = "0.5"
 rand = { workspace = true }
