+++ datafusion/physical-expr/src/string_expressions.rs
@@ -21,7 +21,6 @@
 
 //! String expressions
 
-use std::iter;
 use std::sync::Arc;
 
 use arrow::{
@@ -31,9 +30,7 @@ use arrow::{
     },
     datatypes::DataType,
 };
-use uuid::Uuid;
 
-use datafusion_common::utils::datafusion_strsim;
 use datafusion_common::Result;
 use datafusion_common::{
     cast::{as_generic_string_array, as_int64_array, as_string_array},
@@ -332,193 +329,4 @@ pub fn ends_with<T: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
     let result = arrow::compute::kernels::comparison::ends_with(left, right)?;
 
     Ok(Arc::new(result) as ArrayRef)
-}
-
-/// Prints random (v4) uuid values per row
-/// uuid() = 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11'
-pub fn uuid(args: &[ColumnarValue]) -> Result<ColumnarValue> {
-    let len: usize = match &args[0] {
-        ColumnarValue::Array(array) => array.len(),
-        _ => return exec_err!("Expect uuid function to take no param"),
-    };
-
-    let values = iter::repeat_with(|| Uuid::new_v4().to_string()).take(len);
-    let array = GenericStringArray::<i32>::from_iter_values(values);
-    Ok(ColumnarValue::Array(Arc::new(array)))
-}
-
-/// OVERLAY(string1 PLACING string2 FROM integer FOR integer2)
-/// Replaces a substring of string1 with string2 starting at the integer bit
-/// pgsql overlay('Txxxxas' placing 'hom' from 2 for 4) → Thomas
-/// overlay('Txxxxas' placing 'hom' from 2) -> Thomxas, without for option, str2's len is instead
-pub fn overlay<T: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
-    match args.len() {
-        3 => {
-            let string_array = as_generic_string_array::<T>(&args[0])?;
-            let characters_array = as_generic_string_array::<T>(&args[1])?;
-            let pos_num = as_int64_array(&args[2])?;
-
-            let result = string_array
-                .iter()
-                .zip(characters_array.iter())
-                .zip(pos_num.iter())
-                .map(|((string, characters), start_pos)| {
-                    match (string, characters, start_pos) {
-                        (Some(string), Some(characters), Some(start_pos)) => {
-                            let string_len = string.chars().count();
-                            let characters_len = characters.chars().count();
-                            let replace_len = characters_len as i64;
-                            let mut res =
-                                String::with_capacity(string_len.max(characters_len));
-
-                            //as sql replace index start from 1 while string index start from 0
-                            if start_pos > 1 && start_pos - 1 < string_len as i64 {
-                                let start = (start_pos - 1) as usize;
-                                res.push_str(&string[..start]);
-                            }
-                            res.push_str(characters);
-                            // if start + replace_len - 1 >= string_length, just to string end
-                            if start_pos + replace_len - 1 < string_len as i64 {
-                                let end = (start_pos + replace_len - 1) as usize;
-                                res.push_str(&string[end..]);
-                            }
-                            Ok(Some(res))
-                        }
-                        _ => Ok(None),
-                    }
-                })
-                .collect::<Result<GenericStringArray<T>>>()?;
-            Ok(Arc::new(result) as ArrayRef)
-        }
-        4 => {
-            let string_array = as_generic_string_array::<T>(&args[0])?;
-            let characters_array = as_generic_string_array::<T>(&args[1])?;
-            let pos_num = as_int64_array(&args[2])?;
-            let len_num = as_int64_array(&args[3])?;
-
-            let result = string_array
-                .iter()
-                .zip(characters_array.iter())
-                .zip(pos_num.iter())
-                .zip(len_num.iter())
-                .map(|(((string, characters), start_pos), len)| {
-                    match (string, characters, start_pos, len) {
-                        (Some(string), Some(characters), Some(start_pos), Some(len)) => {
-                            let string_len = string.chars().count();
-                            let characters_len = characters.chars().count();
-                            let replace_len = len.min(string_len as i64);
-                            let mut res =
-                                String::with_capacity(string_len.max(characters_len));
-
-                            //as sql replace index start from 1 while string index start from 0
-                            if start_pos > 1 && start_pos - 1 < string_len as i64 {
-                                let start = (start_pos - 1) as usize;
-                                res.push_str(&string[..start]);
-                            }
-                            res.push_str(characters);
-                            // if start + replace_len - 1 >= string_length, just to string end
-                            if start_pos + replace_len - 1 < string_len as i64 {
-                                let end = (start_pos + replace_len - 1) as usize;
-                                res.push_str(&string[end..]);
-                            }
-                            Ok(Some(res))
-                        }
-                        _ => Ok(None),
-                    }
-                })
-                .collect::<Result<GenericStringArray<T>>>()?;
-            Ok(Arc::new(result) as ArrayRef)
-        }
-        other => {
-            exec_err!("overlay was called with {other} arguments. It requires 3 or 4.")
-        }
-    }
-}
-
-///Returns the Levenshtein distance between the two given strings.
-/// LEVENSHTEIN('kitten', 'sitting') = 3
-pub fn levenshtein<T: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
-    if args.len() != 2 {
-        return exec_err!(
-            "levenshtein function requires two arguments, got {}",
-            args.len()
-        );
-    }
-    let str1_array = as_generic_string_array::<T>(&args[0])?;
-    let str2_array = as_generic_string_array::<T>(&args[1])?;
-    match args[0].data_type() {
-        DataType::Utf8 => {
-            let result = str1_array
-                .iter()
-                .zip(str2_array.iter())
-                .map(|(string1, string2)| match (string1, string2) {
-                    (Some(string1), Some(string2)) => {
-                        Some(datafusion_strsim::levenshtein(string1, string2) as i32)
-                    }
-                    _ => None,
-                })
-                .collect::<Int32Array>();
-            Ok(Arc::new(result) as ArrayRef)
-        }
-        DataType::LargeUtf8 => {
-            let result = str1_array
-                .iter()
-                .zip(str2_array.iter())
-                .map(|(string1, string2)| match (string1, string2) {
-                    (Some(string1), Some(string2)) => {
-                        Some(datafusion_strsim::levenshtein(string1, string2) as i64)
-                    }
-                    _ => None,
-                })
-                .collect::<Int64Array>();
-            Ok(Arc::new(result) as ArrayRef)
-        }
-        other => {
-            exec_err!(
-                "levenshtein was called with {other} datatype arguments. It requires Utf8 or LargeUtf8."
-            )
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use arrow::array::Int32Array;
-    use arrow_array::Int64Array;
-
-    use datafusion_common::cast::as_int32_array;
-
-    use super::*;
-
-    #[test]
-    fn to_overlay() -> Result<()> {
-        let string =
-            Arc::new(StringArray::from(vec!["123", "abcdefg", "xyz", "Txxxxas"]));
-        let replace_string =
-            Arc::new(StringArray::from(vec!["abc", "qwertyasdfg", "ijk", "hom"]));
-        let start = Arc::new(Int64Array::from(vec![4, 1, 1, 2])); // start
-        let end = Arc::new(Int64Array::from(vec![5, 7, 2, 4])); // replace len
-
-        let res = overlay::<i32>(&[string, replace_string, start, end]).unwrap();
-        let result = as_generic_string_array::<i32>(&res).unwrap();
-        let expected = StringArray::from(vec!["abc", "qwertyasdfg", "ijkz", "Thomas"]);
-        assert_eq!(&expected, result);
-
-        Ok(())
-    }
-
-    #[test]
-    fn to_levenshtein() -> Result<()> {
-        let string1_array =
-            Arc::new(StringArray::from(vec!["123", "abc", "xyz", "kitten"]));
-        let string2_array =
-            Arc::new(StringArray::from(vec!["321", "def", "zyx", "sitting"]));
-        let res = levenshtein::<i32>(&[string1_array, string2_array]).unwrap();
-        let result =
-            as_int32_array(&res).expect("failed to initialized function levenshtein");
-        let expected = Int32Array::from(vec![2, 3, 2, 3]);
-        assert_eq!(&expected, result);
-
-        Ok(())
-    }
 }
